Added FM Stereo support
git-svn-id: https://subversion.assembla.com/svn/sdrsharp/trunk@354 d3cde5d6-82b8-4669-8b62-43054d6781d8

+++ Radio/LsbDetector.cs
@@ -21,7 +21,7 @@
 
         public int BfoFrequency
         {
-            get { return _bfo.Frequency; }
+            get { return (int) _bfo.Frequency; }
             set { _bfo.Frequency = value; }
         }
     }
